refactor: main page
